Restructure and document impure tree implementation
