+++ haskey.cabal
@@ -20,10 +20,12 @@ library
   exposed-modules:
     Data.BTree.Alloc.Append
     Data.BTree.Alloc.Class
-    Data.BTree.Delete
-    Data.BTree.Fold
-    Data.BTree.Insert
-    Data.BTree.Lookup
+    Data.BTree.Impure
+    Data.BTree.Impure.Delete
+    Data.BTree.Impure.Fold
+    Data.BTree.Impure.Insert
+    Data.BTree.Impure.Lookup
+    Data.BTree.Impure.Structures
     Data.BTree.Primitives
     Data.BTree.Primitives.Height
     Data.BTree.Primitives.Ids
@@ -33,12 +35,14 @@ library
     Data.BTree.Primitives.Value
     Data.BTree.Pure
     Data.BTree.Pure.TwoThree
-    Data.BTree.Store.Class
     Data.BTree.Store.Binary
+    Data.BTree.Store.Class
     Data.BTree.Store.File
 
   other-modules:
-    Data.BTree.Internal
+    Data.BTree.Utils.List
+    Data.BTree.Utils.Map
+    Data.BTree.Utils.Vector
 
   other-extensions:
     DataKinds
@@ -70,9 +74,9 @@ test-suite haskey-properties
   type:                exitcode-stdio-1.0
   other-modules:
     Properties.Alloc.Append
-    Properties.Fold
-    Properties.Insert
-    Properties.Primitives
+    Properties.Impure.Fold
+    Properties.Impure.Insert
+    Properties.Impure.Structures
     Properties.Primitives.Height
     Properties.Primitives.Ids
     Properties.Primitives.Index
