+++ src/Data/BTree/Utils/List.hs
@@ -1,0 +1,6 @@
+module Data.BTree.Utils.List where
+
+import Data.Maybe (listToMaybe)
+
+safeLast :: [a] -> Maybe a
+safeLast = listToMaybe . reverse
