+++ tests/Properties.hs
@@ -5,9 +5,9 @@ module Main (main) where
 import Test.Framework (Test, defaultMain)
 
 import qualified Properties.Alloc.Append
-import qualified Properties.Fold
-import qualified Properties.Insert
-import qualified Properties.Primitives
+import qualified Properties.Impure.Fold
+import qualified Properties.Impure.Insert
+import qualified Properties.Impure.Structures
 import qualified Properties.Primitives.Height
 import qualified Properties.Primitives.Ids
 import qualified Properties.Primitives.Index
@@ -21,13 +21,13 @@ import qualified Properties.Store.File
 tests :: [Test]
 tests =
     [ Properties.Alloc.Append.tests
-    , Properties.Fold.tests
-    , Properties.Insert.tests
+    , Properties.Impure.Fold.tests
+    , Properties.Impure.Insert.tests
+    , Properties.Impure.Structures.tests
     , Properties.Primitives.Height.tests
     , Properties.Primitives.Ids.tests
     , Properties.Primitives.Index.tests
     , Properties.Primitives.Leaf.tests
-    , Properties.Primitives.tests
     , Properties.Pure.tests
     , Properties.Store.Binary.tests
     , Properties.Store.File.tests
